+++ tiledb/sm/array_schema/array_schema.cc
@@ -291,6 +291,26 @@ const Dimension* ArraySchema::dimension(const std::string& name) const {
   return it == dim_map_.end() ? nullptr : it->second;
 }
 
+std::vector<std::string> ArraySchema::dim_names() const {
+  auto dim_num = this->dim_num();
+  std::vector<std::string> ret;
+  ret.reserve(dim_num);
+  for (uint32_t d = 0; d < dim_num; ++d)
+    ret.emplace_back(domain_->dimension(d)->name());
+
+  return ret;
+}
+
+std::vector<Datatype> ArraySchema::dim_types() const {
+  auto dim_num = this->dim_num();
+  std::vector<Datatype> ret;
+  ret.reserve(dim_num);
+  for (uint32_t d = 0; d < dim_num; ++d)
+    ret.emplace_back(domain_->dimension(d)->type());
+
+  return ret;
+}
+
 unsigned int ArraySchema::dim_num() const {
   return domain_->dim_num();
 }
@@ -298,12 +318,16 @@ unsigned int ArraySchema::dim_num() const {
 void ArraySchema::dump(FILE* out) const {
   if (out == nullptr)
     out = stdout;
-  fprintf(out, "- Array type: %s\n", array_type_str(array_type_).c_str());
-  fprintf(out, "- Cell order: %s\n", layout_str(cell_order_).c_str());
-  fprintf(out, "- Tile order: %s\n", layout_str(tile_order_).c_str());
-  fprintf(out, "- Capacity: %" PRIu64 "\n", capacity_);
-  fprintf(out, "- Allows duplicates: %s\n", (allows_dups_ ? "true" : "false"));
-  fprintf(out, "- Coordinates filters: %u", (unsigned)coords_filters_.size());
+
+  std::stringstream ss;
+  ss << "- Array type: " << array_type_str(array_type_) << "\n";
+  ss << "- Cell order: " << layout_str(cell_order_) << "\n";
+  ss << "- Tile order: " << layout_str(tile_order_) << "\n";
+  ss << "- Capacity: " << capacity_ << "\n";
+  ss << "- Allows duplicates: " << (allows_dups_ ? "true" : "false") << "\n";
+  ss << "- Coordinates filters: " << coords_filters_.size();
+  fprintf(out, "%s", ss.str().c_str());
+
   coords_filters_.dump(out);
   fprintf(
       out,